refactor: move search input components to their own folders
