+++ packages/app-explorer/src/systems/Core/components/Search/SearchInput/index.ts
@@ -1,0 +1,1 @@
+export * from './SearchInput';
