+++ packages/app-explorer/src/systems/Core/components/Search/SearchResultDropdown/index.ts
@@ -1,0 +1,1 @@
+export * from './SearchResultDropdown';
