+++ packages/app-explorer/src/systems/Core/components/Search/styles.ts
@@ -2,34 +2,6 @@ import { tv } from 'tailwind-variants';
 
 export const styles = tv({
   slots: {
-    searchBox: [
-      'transition-all duration-200 [&[data-active=false]]:ease-in [&[data-active=true]]:ease-out',
-      'group justify-center items-center',
-      'block left-0 w-full', // needed for properly execution of transitions
-      '[&[data-active=true]]:w-[calc(100vw+1px)] [&[data-active=true]]:left-[-64px] tablet:[&[data-active=true]]:w-full',
-      '[&[data-active=true]]:absolute tablet:[&[data-active=true]]:left-0 [&[data-active=true]]:right-0',
-      '[&[data-active=true]]:top-[-14px] tablet:[&[data-active=true]]:top-[-4px] desktop:[&[data-active=true]]:top-[-20px]',
-      '[&[data-active=true]]:z-50',
-    ],
-    dropdownItem: 'hover:bg-border focus:bg-border cursor-pointer',
-    inputContainer: 'w-full',
-    inputWrapper: [
-      'outline-none h-[40px] group-[&[data-active=true]]:h-[60px] tablet:group-[&[data-active=true]]:h-[40px]',
-      'group-[&[data-active=true]]:rounded-none tablet:group-[&[data-active=true]]:rounded-[var(--text-field-border-radius)] ',
-      'border-x-[1px] border-y-[1px] group-[&[data-active=true]]:border-x-0 group-[&[data-active=true]]:border-y-0 tablet:group-[&[data-active=true]]:border-x-[1px] tablet:group-[&[data-active=true]]:border-y-[1px]',
-      'border-[var(--color-border)] shadow-none',
-      'bg-none dark:bg-[var(--color-surface)] group-[&[data-active=true]]:bg-[var(--color-panel-solid)]',
-      '[&_.rt-TextFieldChrome]:bg-gray-1 [&_.rt-TextFieldChrome]:outline-none',
-      '[&_.rt-TextFieldChrome]:[&[data-opened=true]]:rounded-b-none',
-      'group-[&[data-active=true]]:[&_.rt-TextFieldChrome]:shadow-none',
-    ],
     searchSize: 'w-full sm:w-[400px] group-[&[data-active=true]]:w-full',
-    inputActionsContainer:
-      '[&[data-show=false]]:hidden absolute flex items-center h-full right-0 top-0 transform',
-    dropdownContent: [
-      'ml-[-10px] tablet:ml-0',
-      'mt-[-10px] rounded-t-none shadow-none border border-t-0 border-border',
-      '[&[data-active=true]]:border-t-0',
-    ],
   },
 });
